+++ CHANGELOG.md
@@ -49,7 +49,7 @@
 
 
 * 1.1.7
-  - added fuzzy search in reporter
+  - added search in reporter
 
 * 1.1.6
   - fixed UI layout in small screen